sched-simple: fix rlist_alloc_nnodes() algorithm
Problem: when the number of nodes is specified in an
allocation request, sched-simple does not always allocate
the requested number of nodes, and may fail requests that
should succeed.

rlist_alloc_nnodes() repeatedly sorts the node list by least
loaded node and picks the first node with enough cores for
a "slot".  However, the first node may be out of cores,
or may have already been used, while another has not been used.

If a node is empty, remove it from consideration.  Otherwise,
force it to the back of the list to ensure all N nodes are
considered at east once.

Fixes #2473 (fix courtesy @grondo)