shell: evlog: add logging plugin to log to output eventlog
Add a shell plugin which subscribes to the 'shell.log' plugin hook
and converts all log messages at or below the current severity level
into RFC 24 Log Events, and appends them to the job's output eventlog.

An "eventlogger" object is used to perform the eventlog appends, so
by default log events will be batched for up to 10ms. FLUX_SHELL_FATAL
log messages are written synchronously so that the log event is
guaranteed to be recorded before the plugin returns and the shell
possibly exits.

Fixes #2319