shell: wait asynchronously in shell_barrier
Problem: If one shell in a multi-shell job hits a job exception
during initialization, the other shells may get stuck in the
shell init barrier for the entire duration of the job-exec module's
kill timeout, instead of aborting immediately.

Replace the currently synchronous barrier with an asynchronous
version so that the shell can watch the job eventlog and abort
the barrier on any job exception.

This should also speed up cancellation of jobs during shell
initialization.