libflux/handle: use config file for flux_open() URI
Problem: when FLUX_URI is not set, Flux utils fall back to
the compiled-in (installed) URI.  This won't work on a login
node that is not running an installed Flux broker.  The URI
to fall back to should be configurable by teh sys admins.

If the URI argument and FLUX_URI environment variable are
unset, try to parse the Flux config.  If successful, cache
it in the handle for future use.  If "flux_uri" is defined
in the config, use that.  if it is not defined or if the
config was not successfully parsed, fall back to the
compiled-in value.