job-ingest: launch `.py` validators with configured python
`validate-schema.py.in` was an autoconf input file with a `#!@PYTHON@`
template at the top of the file.  This ensured that the
`validate-schema.py` script always ran with the python interpreter that
Flux was configured with.  This works as long as the python interpreter
path is short enough.  When the path is > 127 characters, the script
fails to execute due to kernel limits.

Rename `validate-schema.py.in` to `validate-schema.py` and remove the
`#!@PYTHON@` template shebang line.  Update the `job-ingest` module to
launch validators ending with `.py` with the configured python
interpreter.  `job-ingest` executes other scripts as before.

Add a test to ensure that non-python executables can still be launched
as validators.

Fixes #2491