t1004-statwatcher: fix read hang on Ubuntu 19.10
Running this test on Ubuntu 19.10 would always cause a timeout.  For
whatever reason, the second call to the statwatcher callback would not
read any additional data from the newly appended to file.  After
attempting to `strace` and `ltrace` the test, we concluded it is mostly
likely due to the EOF error flag being set on the FILE stream and not
being cleared before additional reads are attempted.

Fix by always open anew the file each time the statwatcher callback is
called.