shell: add exit code parameter to shell_die(), shell_die_errno()
Problem: In some cases, callers of shell_die() or shell_die_errno()
may want the shell to die from a different exit code than 1. Also
this is more standard practice for these kinds of fatal logging
functions.

Change the signature of flux_shell_fatal() to take an exit_code
parameter and add this to the shell_die() and shell_die_errno()
calls.