job-exec: fix use-after-free in job cancellation
Problem: The jobinfo_cancel() function does not protect itself from
being called while a job cancellation is already in progress. If
this occurs, then a new kill_timer overwrites the old timer, and
when the jobinfo structure is destroyed, only the new timer is
stopped, leaving the old timer to fire and access freed memory.

If jobinfo_cancel() check if the kill_timer is already active, and
simply return if so.

Fixes #2519