broker: parse config file(s)
Parse TOML config files, caching result in the broker's
internal flux_t handle for use by broker subsystems.

If the config file has errors, this is fatal to the broker.
If no configuration exists (e.g. *.toml has no match), this
is also fatal, but this should be OK because the boot.toml file
is present in-tree and is also installed.  (It has no effect
unless the 'config' boot method is explicitly selected on
the broker command line).