t1105-proxy.t: eliminate race by switching to event-trace.lua
the racy test was testing that multiple, back-to-back subscriptions only
resulted in a single `subscribe` dmesg.  The test was subscribing to the
timed heartbeat.  If the heartbeat arrived before the second
subscription could be set, then the test would fail will duplicate
subscription messages.

Switch the subscription topic from `hb` to `foobar`.  Switch from using
`flux event sub` to `event-trace.lua`, which ensures that the
subscription is in place before the subcommand (event pub) is run.

Fixes #2522