t0004-event.t: fix racy tests use unique event topics between tests
The test from "rank 0" that precedes the test for events from "rank 3"
use the same event names. If the first test times out because a crazy
person has run `make -j check`, then the second test could possibly
subscribe and recieve the `testcase.eof` event first before the other
events it is expecting.

The fix is to:
- Increase the timeout for these tests to handle constrained case better
- Use different named events between the two tests.

Closes #2517