broker/boot_config: add a bit of inline documentation
Problem: recently it was noted that hostnames could not
be used in the "tbon-endpoints" array when configuring
a flux instance to bootstrap from config files.  This was
not documented.

Most likely we will need to make our configuration more flexible.
However, for now, note this limitation in src/broker/boot_config.h
where the [boostrap] table is described.