modules/kvs: Take reference on used cache entries
When a lookup stall occurs on a missing cache reference, a pointer
to a treeobj is stored in the lookup handle.  This treeobj is
owned by a cache entry and is assumed to be valid when a lookup
resumes.  In other words, the cache entry shouldn't be destroyed.

If the lookup stall is slow, perhaps due to a busy KVS, it is
possible for the cache entry which owns the cached treeobj pointer
to be destroyed via cache expiration.

Whenever the lookup handle internally "saves" a pointer to the
treeobj (or an object within the treeobj), take a reference on the
cache entry, so that it cannot be accidentally expired.

Fixes #2505