modules/kvs: Decrease treeobj validation time
Profiling analysis showed that the KVS can slow down when some
KVS entries, such as output logs, are continually appended to.

Part of the slow down was due to the large amount of time spent
in blobref_validate() checks, which were called via treeobj_validate().
This can be especially slow when the valref treeobjects have a
very long array of blobrefs stored internally, as treeobj_validate()
will verify that every blobref in the array is valid.

To improve performance, call treeobj_insert_entry_novalidate() instead
of treeobj_insert_entry() when appending a blobref to an existing
valref.  All prior entries in the valref treeobj have been checked by
earlier kvs transactions, we only have to check that the newest blobref
is valid.