modules/job-manager: use zhashx_t for active jobs
Problem: the 'queue' is not a great fit for the main
hash of active jobs.  The only "queue" is really in the
alloc module, where jobs are queued in (1) priority,
then (2) submit time order for sending to the scheduler.
But that uses a separate container.

Use a zhashx_t container to store ctx->active_jobs instead
of the queue pseudo-container.

Update unit tests.

Update internal users.  There are three tricky cases:

1) job-manager.list handler

Since job-info.list now supports the primary job listing tools,
this handler need only provide functionality needed by sharness
tests and t/job-manager/list-jobs.c.

Add a first/next interfaces to alloc.c that allow its queue of
pending jobs to be iterated over.

First list pending jobs in order, then dump the hash (excluding
pending jobs) in random order.  This satisfies all the testing
that we currently do.

Drop support for configuring the set of attributes returned in
the response.

2) job-manager.drain handler

queue_set_notify_empty() allowed a callback to be triggered when
the queue transitioned to empty.

Replace that with a callback with a hardwired call to
drain_empty_notify() in event.c, when the last job transitions
to INACTIVE.

If the queue is not being drained, it's a no-op.

3) job-manager.priority handler

This arguably didn't work before.  After updating the priority value,
it would call queue_reorder(), which changed the job's position in
the "queue".  However, the real queue of pending jobs in alloc.c
was unaltered.

Add a reorder interface to alloc.c that alters the job's position
in the queue of pending jobs.  Call this instead.

This still leaves the problem of telling a scheduler that the
pending queue needs to be refreshed as a TBD (#2069).