libjob/job_hash: add job hash helpers
Add helpers for storing user-defined job objects in a zhashx_t
using the integer jobid as hash key.  The key is presumed to be
a "owned" by the job object.

This is for internal flux-core use only since it it presumes
CZMQ is being used, which is not required for Flux API users.