cmd/flux-remote: don't use $TMPDIR for remote socket
Problem: flux-remote uses $TMPDIR instead of /tmp as
the path for the remote socket, if set locally.  But it's
not set remotely on our systems unless the user logs in,
and in fact /var/tmp/$USER is only created the first time
the user logs in.

Use /tmp as the prefix for the remote socket.