shell: input: remove input-ready synchronization
The shell input plugin generates an input-ready event to the guest
exec eventlog in order for tasks to ensure this eventlog exists before
they invoke a KVS watch on it.

However, this syncronization is not necessary, since the KVS watch on
the input eventlog is installed in the task.init callback, in which the
shell is guaranteed to have exited the shell initialization barrier.

Remove the extra synchronization and simply watch the input eventlog
from each task in the task.init callbacks.  # Please enter the commit
message for your changes. Lines starting