cmd/flux-job: Add new stats subcommand
The new stats subcommand will send an RPC to job-info.job-stats to
get current job statistics.  At the moment the job statistics simply
are the number of jobs in each job state and the total number of
jobs.

For the time being, simply output the job stats json object in raw
form the stdout.  No special formatting is done at the moment.