modules/job-manager: capture end event
Add 'end_event' member to struct job, and ensure that the
event that caused the job to transition to FLUX_JOB_CLEANUP
is captured there.  This is either a fatal exception event,
or a finish event in the RFC 21 state machine.

This event, with context, should be sufficient payload for
a "wait" response to allow a caller to determine, at a coarse
level of detail, what happened to a job.