libjob: add flux_job_wait()
Add an API interface for waiting for a job.

Use flux_job_wait() to get a future which is fulfilled
once the job reaches INACTIVE state.

Use flux_job_wait_get_status() to retrieve a boolean
success vaule and textual error message, which details
the exception if a job was terminated that way,
or the exit status of shells.