modules/job-manager: add support for flux_job_wait()
Add support in the job manager for flux_job_wait().

If a job is currently active when a wait request is
received for it, tack the request onto the 'struct job'.

When a job with the FLUX_JOB_WAITABLE flag transitions
to FLUX_JOB_INACTIVE state, there are two cases:
(1) There is a job->waiter - respond to it immediately.
(2) There is no job->waiter - insert the job into the wait->zombies hash.

If a job is not currently active when a wait request is
received for it, find the job in the wait->zombies hash.

If the job id is FLUX_JOBID_ANY, then search first for a
zombie, then for a waitable active job.  If there are no
eligible jobs to wait on, an error is returned.

The rules for users are
- Only the instance owner can set FLUX_JOB_WAITABLE on a job,
  therefore only the instance owner can call flux_job_wait().
- The wait call is destructive to the waitable state, so it
  may only be used once.