shell: input: add function to get pointer to task input struct
Add a function to return the correct `struct shell_task_input *` from
a `struct shell_input *` given a `flux_shell_task_t *`. Simply uses
the task's local index as the index into the task_inputs array.

As a side-effect, this removes the need for the task_inputs_count
member of struct shell_input, so remove it.