modules/kvs: Fix duplicate append corner case
When a kvs transaction stalls and replays, all operations in the
transaction are replayed.  Under most circumstances, this is fine.
If a value has been written to the new root copy before the replay,
it's ok if it is written again.

However, if an append has been written to the new root copy before
the replay, a replay will result in the append being written twice.

In order to avoid this, count how many appends have occurred
within a transaction.  If an append has occurred, reset the root cpy
back to the original so that the replay can start with a fresh
new copy.

Fixes #2547