bindings/python: rename Wrapper._clear()
Allow future.pimpl._clear() to be called from other parts
of the binding by dropping an underscore from its prefix.

The _clear method has some extra protections to ensure the C
destructor is only called once, thus future.pimpl._clear()
is the preferred way to explicitly dispose of a future.

This is a prereq for explicit cleanup of futures in the
job() class.

Originally proposed in #2553.