README.md: update build docs for Python 3
Problem: PYTHON_VERSION must be set to build with Python 3
on modern distros, since /usr/bin/python is usually Python 2,
but this is not documented.

Since the intent is to eventually get rid of Python 2 support,
we want to encourage new users to go with Python 3.  Therefore,
change README.md to list the Python 3 prerequisite packages instead
of python2 ones, and add PYTHON_VERSION=3.6 to the configure line.

Describe how PYTHON_VERSION works, what to do if you really need
Flux to use Python 2, and warn that Python 2 support in Flux
should be considered on its way out.

Fixes #2556

Fix an unrelated spelling error.