libflux/future: take a ref on flux_t handle
Problem: destroying a flux_t before an RPC future causes
an assertion failure in the tagpool, or a segfault.

The future doesn't take a reference on the flux_t handle, so
this destroy order doesn't work.

Have flux_future_set_flux() take a reference on the flux_t
which is dropped when the future is destroyed.

Fixes #2567