libflux/future: add unit test for destroy order
Add a unit test that creates an object that looks similar
to a flux RPC, and verify that the object and the flux_t
can be destroyed in any order.

Try when future is fulfilled in sync and async modes,
and try with future unfulfilled.

Also checked manually under valgrind.