shell: add unpack variants of get info calls to shell.h api
Problem: It is very cumbersome to use the flux_shell_get_info()
and flux_shell_get_rank_info() calls from C plugins. In order to
get one value out, the entire JSON object needs to be parsed, then
discarded. Additionally, a json_t object is created and destroyed
on every call in order to construct the JSON string.

Cache the json_t object for these functions in the shell's aux hash
after first creation, and use the cached object to offer unpack
interfaces for these functions as a convenice.