python/future: fix premature destructon of future's handles
When calling `future.get_flux()`, the returned handle is wrapped in a
`Flux` object.  When this object leaves scope, `flux_close` is called,
closing the handle prematurely.

Increment the reference count of the handle when wrapping it to prevent
this reference counting bug.

Add an explicit test for this bug to `t0012-futures.py`.