broker/boot_config: convert to new config format
Change the [bootstrap] configuration to allow bind URI,
and connect URI to be expressed independently, and to
determine the rank by matching the hostname, not the
local interface addresses.

Rename the rank-ordered array from 'tbon-endpoints' to
'hosts' for clarity.  A hosts entry is a table containing
the keys 'host' (required), 'bind' (optional), and 'connect'
(optional).

The keys 'default_port', 'default_bind', and 'default_connect'
may be defined.  The default bind and connect values are
used when those keys are not explicitly set in the hosts entry.
The 'default_port' value is substituted for the token "%p" if
it occurs in bind or connect strings.  The 'host' value in a
hosts entry entry is substituted for the token "%h" if it
occurs in bind or connect strings.

The 'host' key may included embedded bracketed idsets.

Example 1024 node cluster config:

[bootstrap]

default_port = 8050
default_bind = "tcp://en0:%p"
default_connect = "tcp://e%h:%p"

hosts = [
    {
        host="fluke0",
        bind="tcp://en4:9001",
        connect="tcp://fluke-mgmt:9001"
    },
    { host = "fluke[1-1023]" },
]

In the example, the bind address is "tcp:/en4:9001" on rank 0
and "tcp://en0:8050" on ranks 1-1023.  The connect address for
rank 0 is "tcp://fluke-mgmt:9001"; "tcp://efluke1:8050" for rank 1;
"tcp://efluke2:8050" for rank 2, etc..

Update existing sharness test.

Update installed config file.

Fixes #2572