broker/log: open file only on rank 0
Problem: when the broker is started with the
'log-filename' attribute set on the command line,
the log file is opened from all broker ranks.
This does not match the documentation or the code.

Move initialization of logbuf->rank in logbuf_initialize()
to before logbuf_register_attrs() is called.  Now the test
for rank 0 is only true on rank 0.

Move the logbuf->h initialization too for consistency.

Fixes #2581