broker/boot_pmi: dont bind to URI if no downstream peers
When the bootstrap method is PMI, don't bind to a zeromq socket
if there are no downstream peers.

Change broker_response_sendmsg() to silently discard a response
if it has exhausted all other routing option and tries to send
it downstream.  This is zeromq's default behavior for unroutable
messages sent on a ROUTER socket anyway (drop).