testsuite/t2601-job-shell.t: don't redirect stderr
Problem: t2601-job-shell.t (test 8 - PMI cliques are correct
for 2 ppn) just failed in travis with no useful output.

This test, and a few tests surrounding it, redirect the stderr
of flux job attach to a file, then ignore the file.

Allow flux job attach's stderr to be included in the verbose
test output