flux-job: MPIR support for parallel debugging into attach
Provide MPIR debug interface within flux-job’s attach
command.

Implement MPIR global symbols including MPIR_proctable,
MPIR_proctable_size and MPIR_partial_attach_ok, which
enable a parallel debugger to gather information on each
and every MPI process of the target job and also to allow
the tool to synchronize itself with the MPI processes.

Set up the MPIR symbols on receiving the shell.init event and
follow the MPIR aborting protocol on the complete event.

Assume that the upper layer would launch the job with
the stop-tasks-in-exec option for launch mode. With this
option, job-shells fork the MPI processes and leave them
stopped at their first machine instruction. This then allows
the debugger to attach and start to debug the job from
the pristine state.

Add support for MPIR_partial_attach_ok. This symbol
indicates to the debugger that so-called subset-attach
is supported. With this support, the debugger can only
attach to a subset of the target MPI processes and debug
them without having to worry about the job’s forward
progress.
In launch mode, this means that the system must
continue those unattached processes by itself and
we do this by sending SIGCONT. To be scalable, however,
SIGCONT is broadcast to every process as debuggers
almost always ignore additional SIGCONT signals
raised to attached processes.

Provide support for so-called attach mode. In this mode,
a debugger attaches and debugs an already running job.
This mode requires neither the initial synchronization nor
system-induced SIGCONT support.

Add debug-emulate option to facilitate testing. This sets
MPIR_being_debugged=1 global symbol which is a
precondition to activate the MPIR protocol. It is the very
symbol that the debugger sets externally as soon as it
attaches to the flux-job command.

Add the mpir and libdebugged libraries as the dependencies
to flux-job in Makefile.am.