python: require python3, make python2 optional
I managed to find a way to make the subproject setup work without
polluting the outer project.  The current setup will always require
python >= 3.6, if configure is given the --with-python2 flag, it will
additionally build python2 bindings which require a python version <=
2.8.  The inner python2 bindings intentionally ignore PYTHON and
PYTHON_VERSION, but respect PYTHON2 and PYTHON2_VERSION instead.

This does not yet include a test hookup or a python2 helper subcommand
but those are coming.