python2: replicate test harness to test python2
This is a basic rework to allow the tests to be run against the python-2
bindings.  Doing it highighted an issue with our current approach of
setting `PYTHONPATH`, that it is used by *both* python2 and python3 and
can only point to one or the other version at a time.  For the purpose
of the tests, I reworked subflux.py to inject the correct PYTHONPATH
components at the start of the variable after the flux command added the
wrong ones, but it may be worth thinking about how we want to handle
this in general.  For situations where flux is installed into the same
prefix as the python install, or where the module can be made part of a
virtualenv or the user will set their pythonpath explicitly, this will
not be an issue, but it may be a usability concern for python2 users.