libflux/mrpc: remove mrpc and tests
Problem: the mrpc class in our API was an early attempt
at making RPCs concurrent.  It is discontiguous with
respect to our newer async API's that return futures,
and it has no way to limit its concurrency so is a potential
DoS danger.  In short, we shouldn't use this.

Now that internal users have been updated, remove the class
and its unit tests.

Fixes #2607