modules/job-info: Refactor job state transitions
In the job info module, refactor how the internal data structures
manage the transition between job states.

At the moment the only data that has to be retrieved from the KVS
occurs when the job is submitted, i.e. state DEPEND.  Code was designed
to not allow users to see a job in the DEPEND state until this data
had been retrieved from the KVS.  After this point, it was assumed
that no additional data would be retrieved from the KVS.

During the time that data is retrieved from the KVS, there is a racy
component of the list service, where additional state transitions could
occur.  For example, the job could complete and be in INACTIVE state
before data from the KVS is retrieved.  This is considered acceptable,
since the job-info list service is eventually consistent.

In the future, some data many not be available during state DEPEND.
For example, R will become available during state RUN and another
KVS lookup would be required to lookup R.  In this case, the race
described above would be unnacceptable.  A user could see a job
is in the RUN state, but cannot see the node count for the job.

Therefore code had to be refactored to allow data to be retrieved after
any state transition.  Users will not see that the job has transitioned
until after the data has been retrieved.

This refactoring modifies the internal data structures to store
all state transitions on a 'next_states' list.  Only after data
from the KVS is retrieved for a particular state, will the job
info module show that transition to callers.  In the event that
no data has to be retrieved for a state, the transition to that
state can occur "automatically".