configure/python: request a python3
For some reason AX_PYTHON_DEVEL only checks the first python it finds,
even if you ask it for a 3.x version, so it fails in many cases where a
python is available.  The best workaround I've thought of so far is to
set PYTHON_VERSION to 3 if it is not set by the user.