job-ingest: support arbitrary validator subprocess args
Currently, `--schema` is a hardcoded argument that is always passed to
the validator subprocess, and the value of that argument is checked to
be a valid path.  Future validators may not require the schema
argument, may require additional arguments, or no arguments at all.

Update job-ingest's argument parsing and the validate functions to
support arbitrary arguments.  Remove the hardcoded `--schema` argument
in job-ingest, and then update the hardcoded config defaults to
include `--schema`.