python/job: add a validation function for jobspec
validates the jobspec by attempting to construct a Jobspec.  If no
exceptions are thrown during construction, then it is assumed to be
valid and returns True.

By default, the validation function will read the `version` key in the
jobspec to determine which Jobspec object to instantiate. An optional
`require_version` is included to override this behavior and force a
particular class to be used.