flux-job: attach: ignore SIGTTIN,SIGTTOU
Problem: Under certain circumstances, flux-job attach being sent
into the background causes the program to be stopped without the
ability to resume it by bringing to foreground, nor via SIGCONT.
Specifically, this has been seen to occur when detaching a debugger
like TotalView. The flux-job process is then orphaned (reparented to
init) so is not under purview of shell job control, yet input to the
terminal still registers as stdin activity, resulting in immediate
SIGTTIN. If the process is continued with SIGCONT, it attempts to
read from stdin and gets SIGTTIN again, and so on.

Ignoring SIGTTIN appears to make this situation better. The program
gets errors when reading from stdin when in the background, effectively
ignoring this input.

Ignoring SIGTTOU is done for completeness. This will allow flux-job
attach to write to the terminal from the background, even for ttys
with the TOSTOP output mode set (though this is thought to be rare)

Fixes #2599