flux-job: attach: remove warning about stdin EOF
Remove the warning "stdin EOF cound not be sent" from flux-job
attach.

This warning only occurs if there was no other data sent. In this
case the warning is probably not needed as it implies either the job
has redirected stdin from somewhere else, or the job has already
exited.

In both cases, it probably doesn't matter that the EOF from flux-job
attach was not recieved, and thus the warning is just confusing for
users.