libdebugged: thin library that defines a variable for debuggers
Define the MPIR_being_debugged variable that is set to 1 by
a parallel debugger such as totalview.

Will allow a python-based Flux front-end command, in particular
flux-mini.py to detect whether it is being attached by a debugger.
If so, it will start the parallel job in a stopped state
to synchronize the job state with the debugger.

Also define the MPIR_Breakpoint function because totalview only sets
MPIR_being_debugged=1, when it is present