modules/job-info: Do not always return job info
Alter list service to not always provide some information
for a specific job.

For example, a job that is still waiting to run has not yet
reached the RUN state and the timestamp of `t_run` is not valid.
Previously, a timestamp of 0.0 was always returned, but now we
do not return a value.