cmd/flux-jobs.py: Update for missing fields
Now that job info does not return values for fields that are not yet
legal, flux-jobs must decide what to do in those circumstances.

For timestamp fields that are not yet valid (e.g. 't_inactive' is not
valid until a job is completed), we will use the output of "0.0" when
the field is not yet valid.