flux-mini: add parallel debugger support to flux mini run
Import and use debugged.py to be able to detect whether
flux-mini is being attached and traced by a parallel debugger
such as totalview.

Pass stop-tasks-in-exec to the shell if debugger attach is
detected to cause all of the processes of the launching job
to be stopped as soon as they are forked and exec’ed.
This allows this job to be synchronized with the controlling
debugger at the first instruction.

Add a hidden --debug-emulate option to facilitate testing,
which causes the flux-mini run command to behave exactly
same as when it is being controlled by a debugger.