test: add a test program that waits on SIGCONT
Raise SIGSTOP and wait until SIGCONT is sent to it.

Will be used in a sharness script that tests the
correctness of attaching a debugger to a running job.
This program will make sure that the target parallel
program will not exit until a debugger attaches
to it and unlocks it.