tessuite: add t2611-debug-emulate.t
Use the --emulate-debug option of flux-job attach to test
two main debug modes: launch and attach modes.

The launch mode is where the target parallel program
is launched under debugger control. In this mode, it will
be spawned  in a stop state so that the debugger will
be able to attach to it before user code is being executed.

The Attach mode is where the target program is already
running and the debugger will attach to and debug it.
Therefore, this mode does not require an explicit
synchronization between the target program and
the debugger.

The added tests check the correctness of these modes
for various interfaces:
- A job launched via flux-mini run interface;
- A job submitted via flux-mini submit interface;
- A job submitted via flux-job submit interface.

It also has a test that ensures a parallel program
launched or submitted with the stop-tasks-in-exec shell option
can be unlocked either by a subsequent debugger attach (in
emulation mode, of course) or an explicit flux-job kill
SIGCONT.