python: remove invalid try/catch in flux.util.help_formatter
As caught by lgtm, the try/catch in flux.util.help_formatter
wasn't catching a valid exception, since the only exception
thrown at that point was from creation of the lambda.

The intended purpose of that code was to safely fall back
to the default argparse.HelpFormatter in the case that some
required internal methods changed. However, any exception
thrown won't be until the class is actually instantiated
and used in parse_args(), so this code wasn't functional
anyway.

Instead, just explicitly check for existence of the
required methods, since there are only 3, and fall back
if any aren't present. (argument changes, however, will
still throw an exception)