libflux/message: add message cred helpers
Problem: boiler plate for checking request message authorization involves
defining rolemask and userid individually, fetching them from the request
with two accessors, and then implementing logic that is identical for
most services.  It is simple, but there is a high consequence of error.

To reduce size of boiler plate and reduce chance of mistakes:
- add 'struct flux_msg_cred' to contain rolemask, userid
- add flux_msg_get_cred(), flux_msg_set_cred() accessors
  that operate on the cred
- add flux_msg_authorize() and flux_msg_cred_authorize()
  to authorize an action if the (message) cred userid matches
  a supplied one, or the message cred rolemask includes OWNER.

Add unit tests.