test: libpmi: avoid unsafe setenv in tests
Problem: libev calls getenv() during initialization which is only
thread-safe if the environment is not being concurrently modified.
In the libpmi tests test/canonical.c and test/canonical2.c, a test
pmi server thread is spawned and *then* the environment is modified
with setenv(), opening a potential data race.

Fix the unsafe setenv() by modifying the environment before the
server thread is started. Unfortunately, this requires moving the
set of PMI_FD into pmi_server_create(), since the PMI client fd is
not otherwise known until the function returns.

Fixes #2668