testsuite: fix tests when /tmp is a link
Problem: If /tmp is a link, e.g. to /var/tmp, a few tests in the
testuite fail because the shell builtin `pwd` and /bin/pwd
(or /usr/bin/pwd) do not report the same thing. E.g.

 cd /tmp && pwd

and

 cd /tmp && /bin/pwd

will report different results, and in the second case, the
result is not explicitly "/tmp".

Fix the very few tests that use /tmp to use the explicit result
of `(cd /tmp && $(which pwd))` where possible.

Fixes #2663