rc3: do not run with bash -e option
Problem: If any step in rc3 fails, the shell immediately exits,
bypassing the rest of the shutdown. At best this results in a slew
of errors about "module was not cleanly shutdown", but at worst
could result in data loss as persistence rules could be skipped.

Remove the -e option from the rc3 shebang line so that the rc3
script produces errors but attempts to run to completion.

Fixes #2672