modules/job-manager: cancel alloc in CLEANUP
Problem: a job in SCHED state that is canceled at about the
same time as the job manager sends its alloc request to the
scheduler may get stuck in CLEANUP state until the scheduler
allocates resources to the job.

As described in #2279, the event message could overtake the
the alloc request on the way to the scheduler.  Now that wehave
a way to cancel alloc requests with an RPC, use that instead.

Partially addresses #2279.

Also [cleanup]:
In event_job_action(), test free_pending flag explicitly before calling
alloc_send_free_request().  It's checked implicitly in the function,
but to be clear and consistent it should be checked at the call site.