shell: wait for input header creation synchronously
Problem: The creation of the guest.input eventlog header is
done asynchronously in the input plugin initialization, which
means that eventlog creation could race with tasks watching
the eventlog. This results in job failure with

 flux_job_event_watch_get: No such file or director

Until we have a way to do asynchronous initialization in the
shell, the input plugin should synchronously ensure the eventlog
has been created before returning from the shell.init callback.
This ensures initialization is complete before entering the
shell.init barrier.

Fixes #2686