cmd/flux-job: fix missing log messages after init failures
Problem: If the job shell fails during initialization, flux-job attach
will ignore all log messages since the shell.init event was never
emitted to the exec eventlog.

To workaround this issue, start watching the output eventlog when the
"complete" event is emitted to the guest.exec.eventlog, if the output
is not already being watched.

Fixes #2490