shell: do not require flux.posix in mvapich.lua
Problem: The shell mvapich.lua plugin requires 'flux.posix' which
is only meant to be called from utilities, since it prints an
error message to stderr and exits. As a result, if the Lua posix
module isn't found, the mvapich.lua plugin will exit out from
under the shell, resulting in a failed job with no error messages.

Do away with the posix module usage in mvapich.lua to avoid this
result, since only dirname() was used anyway.

Fixes #2698