job-ingest: correct validator worker name in logs
Problem: The worker "name" for job-ingest validator executables
is set from the basename of argv0. However, for python validator
scripts, PYTHON_INTERPRETER is substituted for argv0, so all
validator workers are named "python" or "python3.6" or somesuch.

Allow the validator name to be explicitly set in worker_create(),
and always pass validate_path so the python interpreter isn't
inadvertently included as the name.