job-ingest: kill off workers on module unload
Problem: When job-ingest module is unloaded it calls worker_stop()
to terminate any active validation workers. However, this function
stops workers by calling flux_subprocess_close() on stdin, which
only works if the reactor is active. Since the module is exiting
this will never be the case.

Add a function worker_kill() and call this on all workers during
job-ingest module unload. To increase the chances that the signal
is delivered, wait for up to 100ms for a response from the kill
RPC to ensure the broker called the kill(2) system call on any
active pids.

Fixes #2616