libflux: allow NULL name for flux_future_push(3)
Problem: it is annoying to require the children of composite
futures to have an explicit name in flux_future_push(3), when
many times a set of futures are indistinct. The caller is then
required to use sprintf(3) or similar to generate unique names,
reducing the clarity and usefulness of the composite future
wait_any/wait_all interface.

Allow NULL to be passed as the name parameter in flux_future_push(3),
which will generate a unique name for each child future on the fly
using an internal sequence number.